Made some corrections

+++ Backend/src/controller/post.js
@@ -2,7 +2,9 @@ import createHttpError from "http-errors";
 import Post from "../models/post.js";
 
 export const createPost = async (req, res, next) => {
-  const { title, content, author } = req.body;
+  const { title, content } = req.body;
+
+  const author = req.user._id;
 
   try {
     if (!title || !content || !author) {
@@ -25,10 +27,22 @@ export const createPost = async (req, res, next) => {
   }
 };
 
-export const getAllPosts = async (req, res, next) => {
+export const getAllPosts = async (req, res) => {
+  const posts = await Post.find().populate("author", "name email");
+  res.status(200).json({ success: true, posts });
+};
+
+export const getOnePost = async (req, res, next) => {
+  const { id } = req.params;
+  if (!id) {
+    return next(createHttpError(400, "Post id is required"));
+  }
   try {
-    const posts = await Post.find();
-    res.status(200).json({ success: true, posts });
+    const post = await Post.findById(id).populate("author", "name, email");
+
+    if (!post) return next(createHttpError(404, "Post not found"));
+
+    res.status(200).json({ post });
   } catch (error) {
     next(error);
   }
@@ -38,6 +52,9 @@ export const updatePost = async (req, res, next) => {
   const { id } = req.params;
 
   try {
+    if (!id) {
+      return next(createHttpError(400, "Post id is required"));
+    }
     const updatedPost = await Post.findByIdAndUpdate(id, req.body, {
       new: true,
       runValidators: true,
@@ -45,6 +62,11 @@ export const updatePost = async (req, res, next) => {
 
     if (!updatedPost) return next(createHttpError(404, "Post not found"));
 
+    if (req.user.role !== "admin" && updatePost.author !== req.user._id) {
+      return next(
+        createHttpError(403, "Forbidden: you can only update you own post")
+      );
+    }
     res
       .status(200)
       .json({ success: true, message: "Post updated", post: updatedPost });
@@ -57,9 +79,18 @@ export const deletePost = async (req, res, next) => {
   const { id } = req.params;
 
   try {
+    if (!id) {
+      return next(createHttpError(400, "Post id is required"));
+    }
     const deletedPost = await Post.findByIdAndDelete(id);
 
     if (!deletedPost) return next(createHttpError(404, "Post not found"));
+
+    if (req.user.role !== "admin" && updatePost.author !== req.user._id) {
+      return next(
+        createHttpError(403, "Forbidden: you can only delete you own post")
+      );
+    }
 
     res.status(200).json({
       success: true,
